requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
flask==3.1.0
idna==3.10
lxml==5.4.0
python-dotenv==1.1.0
requests==2.32.4
schedule==1.2.2
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
werkzeug==3.1.3
